# api/openapi.json
{"openapi":"3.1.0","info":{"title":"Live Tram API","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/areas":{"get":{"tags":["areas"],"summary":"List all configured areas","operationId":"list_areas","responses":{"200":{"description":"List of all areas","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AreaListResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/areas/{id}":{"get":{"tags":["areas"],"summary":"Get a specific area by ID","operationId":"get_area","parameters":[{"name":"id","in":"path","description":"Area ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Area details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Area"}}}},"404":{"description":"Area not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/areas/{id}/stats":{"get":{"tags":["areas"],"summary":"Get statistics for an area","operationId":"get_area_stats","parameters":[{"name":"id","in":"path","description":"Area ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Area statistics","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AreaStats"}}}},"404":{"description":"Area not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/departures":{"get":{"tags":["departures"],"summary":"List all departures across all stops","operationId":"list_departures","responses":{"200":{"description":"List of all departures","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartureListResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/departures/by-stop":{"post":{"tags":["departures"],"summary":"Get departures for a specific stop by IFOPT ID","operationId":"get_departures_by_stop","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StopDeparturesRequest"}}},"required":true},"responses":{"200":{"description":"Departures for the stop","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StopDeparturesResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/issues":{"get":{"tags":["issues"],"summary":"List all OSM data quality issues","operationId":"list_issues","responses":{"200":{"description":"List of OSM data quality issues","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IssueListResponse"}}}}}}},"/api/routes":{"get":{"tags":["routes"],"summary":"List all routes, optionally filtered by area or type","operationId":"list_routes","parameters":[{"name":"area_id","in":"query","description":"Filter by area ID","required":false,"schema":{"type":["integer","null"],"format":"int64"}},{"name":"route_type","in":"query","description":"Filter by route type (e.g., \"tram\", \"bus\")","required":false,"schema":{"type":["string","null"]}}],"responses":{"200":{"description":"List of routes","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteListResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/routes/{route_id}":{"get":{"tags":["routes"],"summary":"Get a single route with its stops","operationId":"get_route","parameters":[{"name":"route_id","in":"path","description":"Route OSM ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Route details with stops","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteDetail"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/routes/{route_id}/geometry":{"get":{"tags":["routes"],"summary":"Get the geometry of a route as line segments","operationId":"get_route_geometry","parameters":[{"name":"route_id","in":"path","description":"Route OSM ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Route geometry","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteGeometry"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/stations":{"get":{"tags":["stations"],"summary":"List all stations that have platforms linked to them, optionally filtered by area","operationId":"list_stations","parameters":[{"name":"area_id","in":"query","description":"Filter by area ID","required":false,"schema":{"type":["integer","null"],"format":"int64"}}],"responses":{"200":{"description":"List of stations with their platforms and stop positions","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StationListResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/vehicles/by-route":{"post":{"tags":["vehicles"],"summary":"Get all vehicles currently on a route with their stop sequences","operationId":"get_vehicles_by_route","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiclesByRouteRequest"}}},"required":true},"responses":{"200":{"description":"List of vehicles on the route","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiclesByRouteResponse"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"Area":{"type":"object","required":["id","name","south","west","north","east","created_at"],"properties":{"created_at":{"type":"string"},"east":{"type":"number","format":"double"},"id":{"type":"integer","format":"int64"},"last_synced_at":{"type":["string","null"]},"name":{"type":"string"},"north":{"type":"number","format":"double"},"south":{"type":"number","format":"double"},"west":{"type":"number","format":"double"}}},"AreaListResponse":{"type":"object","required":["areas"],"properties":{"areas":{"type":"array","items":{"$ref":"#/components/schemas/Area"}}}},"AreaStats":{"type":"object","required":["area_id","area_name","station_count","platform_count","stop_position_count","route_count"],"properties":{"area_id":{"type":"integer","format":"int64"},"area_name":{"type":"string"},"platform_count":{"type":"integer","format":"int64"},"route_count":{"type":"integer","format":"int64"},"station_count":{"type":"integer","format":"int64"},"stop_position_count":{"type":"integer","format":"int64"}}},"Departure":{"type":"object","description":"A stop event (departure or arrival)","required":["stop_ifopt","event_type","line_number","destination","planned_time"],"properties":{"delay_minutes":{"type":["integer","null"],"format":"int32"},"destination":{"type":"string","description":"For departures: destination; for arrivals: origin"},"destination_id":{"type":["string","null"],"description":"Destination stop ID (for departures) or origin stop ID (for arrivals)"},"estimated_time":{"type":["string","null"]},"event_type":{"$ref":"#/components/schemas/EventType"},"line_number":{"type":"string"},"planned_time":{"type":"string"},"platform":{"type":["string","null"]},"stop_ifopt":{"type":"string"},"trip_id":{"type":["string","null"],"description":"Unique trip identifier (GTFS trip_id) - consistent across all stops for a journey"}}},"DepartureListResponse":{"type":"object","required":["departures"],"properties":{"departures":{"type":"array","items":{"$ref":"#/components/schemas/Departure"}}}},"ErrorResponse":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},"EventType":{"type":"string","description":"Type of stop event","enum":["departure","arrival"]},"IssueListResponse":{"type":"object","required":["issues","count"],"properties":{"count":{"type":"integer","minimum":0},"issues":{"type":"array","items":{"$ref":"#/components/schemas/OsmIssue"}}}},"OsmIssue":{"type":"object","description":"An OSM data quality issue detected during sync","required":["osm_id","osm_type","element_type","issue_type","transport_type","description","osm_url","detected_at"],"properties":{"description":{"type":"string"},"detected_at":{"type":"string"},"element_type":{"type":"string"},"issue_type":{"$ref":"#/components/schemas/OsmIssueType"},"lat":{"type":["number","null"],"format":"double"},"lon":{"type":["number","null"],"format":"double"},"name":{"type":["string","null"]},"osm_id":{"type":"integer","format":"int64"},"osm_type":{"type":"string"},"osm_url":{"type":"string"},"ref":{"type":["string","null"],"description":"The ref tag value (e.g., platform letter \"a\", \"b\")"},"suggested_ifopt":{"type":["string","null"],"description":"Suggested IFOPT (for missing_ifopt issues)"},"suggested_ifopt_distance":{"type":["integer","null"],"format":"int32","description":"Distance in meters to the suggested stop","minimum":0},"suggested_ifopt_name":{"type":["string","null"],"description":"Name of the stop that was matched"},"transport_type":{"$ref":"#/components/schemas/TransportType"}}},"OsmIssueType":{"type":"string","description":"Types of OSM data quality issues","enum":["missing_ifopt","missing_coordinates","orphaned_element","missing_route_ref","missing_name","missing_stop_position","missing_platform"]},"Route":{"type":"object","required":["osm_id","osm_type","route_type"],"properties":{"area_id":{"type":["integer","null"],"format":"int64"},"color":{"type":["string","null"]},"name":{"type":["string","null"]},"network":{"type":["string","null"]},"operator":{"type":["string","null"]},"osm_id":{"type":"integer","format":"int64"},"osm_type":{"type":"string"},"ref":{"type":["string","null"]},"route_type":{"type":"string"}}},"RouteDetail":{"allOf":[{"$ref":"#/components/schemas/Route"},{"type":"object","required":["stops"],"properties":{"stops":{"type":"array","items":{"$ref":"#/components/schemas/RouteStop"}}}}]},"RouteGeometry":{"type":"object","required":["route_id","segments"],"properties":{"route_id":{"type":"integer","format":"int64"},"segments":{"type":"array","items":{"type":"array","items":{"type":"array","items":{"type":"number","format":"double"}}}}}},"RouteListResponse":{"type":"object","required":["routes"],"properties":{"routes":{"type":"array","items":{"$ref":"#/components/schemas/Route"}}}},"RouteStop":{"type":"object","required":["sequence"],"properties":{"platform_id":{"type":["integer","null"],"format":"int64"},"role":{"type":["string","null"]},"sequence":{"type":"integer","format":"int64"},"station_id":{"type":["integer","null"],"format":"int64"},"station_name":{"type":["string","null"]},"stop_position_id":{"type":["integer","null"],"format":"int64"}}},"Station":{"type":"object","required":["osm_id","osm_type","lat","lon","platforms","stop_positions"],"properties":{"area_id":{"type":["integer","null"],"format":"int64"},"lat":{"type":"number","format":"double"},"lon":{"type":"number","format":"double"},"name":{"type":["string","null"]},"osm_id":{"type":"integer","format":"int64"},"osm_type":{"type":"string"},"platforms":{"type":"array","items":{"$ref":"#/components/schemas/StationPlatform"}},"ref_ifopt":{"type":["string","null"]},"stop_positions":{"type":"array","items":{"$ref":"#/components/schemas/StationStopPosition"}}}},"StationListResponse":{"type":"object","required":["stations"],"properties":{"stations":{"type":"array","items":{"$ref":"#/components/schemas/Station"}}}},"StationPlatform":{"type":"object","description":"Platform info nested in station response","required":["osm_id","lat","lon"],"properties":{"lat":{"type":"number","format":"double"},"lon":{"type":"number","format":"double"},"name":{"type":["string","null"]},"osm_id":{"type":"integer","format":"int64"},"ref":{"type":["string","null"]},"ref_ifopt":{"type":["string","null"]}}},"StationStopPosition":{"type":"object","description":"Stop position info nested in station response","required":["osm_id","lat","lon"],"properties":{"lat":{"type":"number","format":"double"},"lon":{"type":"number","format":"double"},"name":{"type":["string","null"]},"osm_id":{"type":"integer","format":"int64"},"platform_id":{"type":["integer","null"],"format":"int64"},"ref":{"type":["string","null"]},"ref_ifopt":{"type":["string","null"]}}},"StopDeparturesRequest":{"type":"object","required":["stop_ifopt"],"properties":{"reference_time":{"type":["string","null"],"description":"Optional reference time (ISO 8601/RFC 3339) for time simulation.\nWhen provided, departures are computed from the static GTFS schedule\naround this time instead of using live real-time data."},"stop_ifopt":{"type":"string"}}},"StopDeparturesResponse":{"type":"object","required":["stop_ifopt","departures"],"properties":{"departures":{"type":"array","items":{"$ref":"#/components/schemas/Departure"}},"stop_ifopt":{"type":"string"}}},"TransportType":{"type":"string","description":"Transport type for both configuration and runtime detection","enum":["tram","bus","subway","train","ferry","unknown"]},"Vehicle":{"type":"object","required":["trip_id","line_number","destination","stops"],"properties":{"destination":{"type":"string","description":"Final destination of this vehicle"},"line_number":{"type":"string","description":"Line number (e.g., \"1\", \"2\", \"3\")"},"origin":{"type":["string","null"],"description":"Origin of this vehicle's journey"},"stops":{"type":"array","items":{"$ref":"#/components/schemas/VehicleStop"},"description":"All stops this vehicle will visit, in order"},"trip_id":{"type":"string","description":"Unique trip identifier (GTFS trip_id)"}}},"VehicleStop":{"type":"object","required":["stop_ifopt","sequence","lat","lon"],"properties":{"arrival_time":{"type":["string","null"],"description":"Arrival time at this stop (ISO 8601)"},"arrival_time_estimated":{"type":["string","null"],"description":"Estimated arrival time (real-time, if available)"},"delay_minutes":{"type":["integer","null"],"format":"int32","description":"Delay in minutes (positive = late, negative = early)"},"departure_time":{"type":["string","null"],"description":"Departure time from this stop (ISO 8601)"},"departure_time_estimated":{"type":["string","null"],"description":"Estimated departure time (real-time, if available)"},"lat":{"type":"number","format":"double","description":"Latitude"},"lon":{"type":"number","format":"double","description":"Longitude"},"sequence":{"type":"integer","format":"int64","description":"Sequence number on the route"},"stop_ifopt":{"type":"string","description":"Stop IFOPT identifier"},"stop_name":{"type":["string","null"],"description":"Stop name (if available)"}}},"VehiclesByRouteRequest":{"type":"object","required":["route_id"],"properties":{"reference_time":{"type":["string","null"],"description":"Optional reference time (ISO 8601/RFC 3339) for time simulation.\nWhen provided, departures are computed from the static GTFS schedule."},"route_id":{"type":"integer","format":"int64","description":"The OSM route ID to get vehicles for"}}},"VehiclesByRouteResponse":{"type":"object","required":["route_id","vehicles"],"properties":{"line_number":{"type":["string","null"]},"route_id":{"type":"integer","format":"int64"},"vehicles":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}}}}}},"tags":[{"name":"areas","description":"Area management endpoints"},{"name":"routes","description":"Route endpoints"},{"name":"stations","description":"Station and platform endpoints"},{"name":"departures","description":"Real-time departure information"},{"name":"vehicles","description":"Live vehicle tracking"},{"name":"issues","description":"OSM data quality issues"}]}