# api/src/sync/types.rs
//! Type definitions for the sync module.

use serde::{Deserialize, Serialize};
use std::collections::HashMap;
use std::sync::Arc;
use tokio::sync::{broadcast, RwLock};
use utoipa::ToSchema;


/// Type of stop event
#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
#[serde(rename_all = "lowercase")]
pub enum EventType {
    Departure,
    Arrival,
}

/// A stop event (departure or arrival)
#[derive(Debug, Clone, Serialize, ToSchema)]
pub struct Departure {
    pub stop_ifopt: String,
    pub event_type: EventType,
    pub line_number: String,
    /// For departures: destination; for arrivals: origin
    pub destination: String,
    /// Destination stop ID (for departures) or origin stop ID (for arrivals)
    pub destination_id: Option<String>,
    pub planned_time: String,
    pub estimated_time: Option<String>,
    pub delay_minutes: Option<i32>,
    pub platform: Option<String>,
    /// Unique trip identifier (GTFS trip_id) - consistent across all stops for a journey
    pub trip_id: Option<String>,
}

impl Departure {
    pub fn planned_departure(&self) -> &str {
        &self.planned_time
    }

    pub fn estimated_departure(&self) -> Option<&str> {
        self.estimated_time.as_deref()
    }
}

/// In-memory store for departure data
pub type DepartureStore = Arc<RwLock<HashMap<String, Vec<Departure>>>>;

/// Shared reference to the GTFS schedule for computing departures at arbitrary times
pub type ScheduleStore = Arc<RwLock<Option<crate::providers::timetables::gtfs::static_data::GtfsSchedule>>>;

/// Update notification for vehicle data changes
#[derive(Debug, Clone, Serialize)]
pub struct VehicleUpdate {
    /// Timestamp when this update was generated
    pub timestamp: String,
    /// Whether this is the initial snapshot or an incremental update
    pub is_initial: bool,
}

/// Sender for vehicle update notifications
pub type VehicleUpdateSender = broadcast::Sender<VehicleUpdate>;